requirements.txt
flask
requests
python-dotenv
gunicorn
flask-sqlalchemy
flask-login
werkzeug
